Fix a small compiler warning.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@3451 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/frontends/xforms/FormMathsSpace.h
@@ -48,7 +48,7 @@ private:
 	boost::scoped_ptr<FD_form_maths_space> dialog_;
 
 	/// The current choice.
-	short space_;
+	int space_;
 };
 
 #endif //  FORM_MATHSSPACE_H
